requirements.txt
fastapi
uvicorn
python-multipart
langchain-google-genai
langchain-groq
python-dotenv